+++ library/publish/library/Github/GITHUB_CONFIG.mjs
@@ -1,1 +1,0 @@
-export const GITHUB_CONFIG_TOKEN_KEY = "github-token";