+++ library/publish/library/Publish/UploadAssetToGiteaRelease.mjs
@@ -1,14 +1,14 @@
 import { blob } from "node:stream/consumers";
 import { createReadStream } from "node:fs";
-import { GetGihubConfig } from "../Github/GetGihubConfig.mjs";
+import { GetGiteaConfig } from "../Gitea/GetGiteaConfig.mjs";
 import { GetReleaseTag } from "./GetReleaseTag.mjs";
 import MIME_DB from "mime-db/db.json" with { type: "json" };
 import { stat } from "node:fs/promises";
 import { basename, extname, join } from "node:path";
 
-export class UploadAssetToGithubRelease {
+export class UploadAssetToGiteaRelease {
     /**
-     * @returns {Promise<UploadAssetToGithubRelease>}
+     * @returns {Promise<UploadAssetToGiteaRelease>}
      */
     static async new() {
         return new this();
@@ -27,7 +27,7 @@ export class UploadAssetToGithubRelease {
      * @param {string | null} asset_name
      * @returns {Promise<void>}
      */
-    async uploadAssetToGithubRelease(path, asset_path, asset_name = null) {
+    async uploadAssetToGiteaRelease(path, asset_path, asset_name = null) {
         const _asset_name = asset_name ?? basename(asset_path);
 
         const _asset_path = join(path, asset_path);
@@ -37,39 +37,38 @@ export class UploadAssetToGithubRelease {
                 path
             );
 
-        console.log(`Upload asset ${asset_path} as ${_asset_name} to github release ${tag}`);
+        console.log(`Upload asset ${asset_path} as ${_asset_name} to gitea release ${tag}`);
 
-        const github_config = await (await GetGihubConfig.new())
-            .getGithubConfig(
+        const gitea_config = await (await GetGiteaConfig.new())
+            .getGiteaConfig(
                 path
             );
 
-        const response = await fetch(`https://api.github.com/repos/${github_config.repository}/releases/tags/${tag}`, {
+        const response = await fetch(`${gitea_config.host}/api/v1/repos/${gitea_config.repository}/releases/tags/${tag}`, {
             headers: {
-                Accept: "application/vnd.github+json",
-                Authorization: github_config.authorization,
-                "User-Agent": github_config["user-agent"]
-            }
+                Authorization: gitea_config.authorization
+            },
+            ...gitea_config["https-certificate-options"]
         });
 
         if (!response.ok || !(response.headers.get("Content-Type")?.includes("application/json") ?? false)) {
             throw response;
         }
 
-        const url = new URL(`https://uploads.github.com/repos/${github_config.repository}/releases/${(await response.json()).id}/assets`);
+        const url = new URL(`${gitea_config.host}/api/v1/repos/${gitea_config.repository}/releases/${(await response.json()).id}/assets`);
         url.searchParams.set("name", _asset_name);
 
         const _response = await fetch(`${url}`, {
             body: await blob(createReadStream(_asset_path)),
             headers: {
-                Authorization: github_config.authorization,
+                Authorization: gitea_config.authorization,
                 "Content-Length": (await stat(_asset_path)).size,
                 "Content-Type": this.#getMimeType(
                     _asset_path
-                ),
-                "User-Agent": github_config["user-agent"]
+                )
             },
-            method: "POST"
+            method: "POST",
+            ...gitea_config["https-certificate-options"]
         });
 
         if (!_response.ok) {